+++ tests/run-all.sh
@@ -1,7 +1,9 @@
 #! /bin/bash
-set -eu
+set -u
 
-E2E_REGEX=${E2E_TEST_REGEX:-*.test.ts}
+# TODO - Remove TypeScript regex after all tests have been migrated to Go.
+E2E_REGEX_GO="./scalers*${E2E_TEST_REGEX:-*_test.go}"
+E2E_REGEX_TS="./scalers*${E2E_TEST_REGEX:-*.test.ts}"
 
 DIR=$(dirname "$0")
 cd $DIR
@@ -15,21 +17,29 @@ counter=0
 executed_count=0
 
 function run_setup {
-    ./node_modules/.bin/ava setup.test.ts
+    go test -v -tags e2e setup_test.go
 }
 
 function run_tests {
     counter=0
     # randomize tests order using shuf
-    for test_case in $(find scalers -name "$E2E_REGEX" | shuf)
+    # TODO - Remove TypeScript regex after all tests have been migrated to Go.
+    for test_case in $(find . -wholename "$E2E_REGEX_GO" -o -wholename "$E2E_REGEX_TS" | shuf)
     do
-        if [[ $test_case != *.test.ts ]] # Skip helper files
+        if [[ $test_case != *_test.go && $test_case != *.test.ts ]] # Skip helper files
         then
             continue
         fi
 
         counter=$((counter+1))
-        ./node_modules/.bin/ava $test_case > "${test_case}.log" 2>&1 &
+        # TODO - Remove condition after all tests have been migrated to Go.
+        if [[ $test_case == *_test.go ]]
+        then
+            go test -v -tags e2e $test_case > "${test_case}.log" 2>&1 &
+        else
+            ./node_modules/.bin/ava $test_case > "${test_case}.log" 2>&1 &
+        fi
+
         pid=$!
         echo "Running $test_case with pid: $pid"
         pids+=($pid)
@@ -62,7 +72,14 @@ function run_tests {
         for test_case in "${retry_lookup[@]}"
         do
             counter=$((counter+1))
-            ./node_modules/.bin/ava $test_case > "${test_case}.retry.log" 2>&1 &
+            # TODO - Remove condition after all tests have been migrated to Go.
+            if [[ $test_case == *_test.go ]]
+            then
+                go test -v -tags e2e $test_case > "${test_case}.retry.log" 2>&1 &
+            else
+                ./node_modules/.bin/ava $test_case > "${test_case}.retry.log" 2>&1 &
+            fi
+
             pid=$!
             echo "Rerunning $test_case with pid: $pid"
             pids+=($pid)
@@ -94,7 +111,7 @@ function wait_for_jobs {
 }
 
 function print_logs {
-    for test_log in $(find scalers -name "*.log")
+    for test_log in $(find . -name "*.log")
     do
         echo ">>> $test_log <<<"
         cat $test_log
@@ -114,7 +131,7 @@ function print_logs {
 }
 
 function run_cleanup {
-    ./node_modules/.bin/ava cleanup.test.ts
+    go test -v -tags e2e cleanup_test.go
 }
 
 function print_failed {